Introduce a mix greecex.zip task
We can use this task to zip the contents of this repository excluding
various folders and files that are not necessary. The reason we might
want a zip file with the contents of this repository is to share it as
context with various LLMs as we work to implement new features.